refactor: 컴포넌트를 page 단위로 분리하도록 구조를 수정
