+++ src/App.tsx
@@ -1,5 +1,5 @@
 import './App.css';
-import CardPage from './components/CardPage/CardPage';
+import CardPage from './pages/CardPage/CardPage';
 
 function App() {
   return <CardPage />;
