+++ src/pages/CardPage/components/CardNumberInput/CardNumberInput.tsx
@@ -1,9 +1,9 @@
 import styled from '@emotion/styled';
-import Input from '../Input/Input';
-import { HandleInputParams } from '../CardPage/CardPage';
-import HelperText from '../HelperText/HelperText';
-import { inputValidation } from '../../validators/inputValidator';
-import useInputValidation from '../../hooks/useInputValidation';
+import Input from '../../../../components/Input/Input';
+import HelperText from '../../../../components/HelperText/HelperText';
+import { inputValidation } from '../../../../validators/inputValidator';
+import useInputValidation from '../../../../hooks/useInputValidation';
+import { HandleInputParams } from '../../CardPage';
 
 type CardNumberInputProps = {
   values: string[];
