+++ src/pages/CardPage/components/PreviewCard/PreviewCard.tsx
